+++ ssh2/agent.pyx
@@ -16,9 +16,34 @@
 
 from pkey cimport PublicKey, PyPublicKey
 from exceptions cimport AgentConnectionError, AgentListIdentitiesError, \
-    AgentGetIdentityError, AgentAuthenticationError
+    AgentGetIdentityError, AgentAuthenticationError, AgentError
 from utils cimport to_bytes
 cimport c_ssh2
+
+
+cdef int agent_auth(char * _username,
+                    c_ssh2.LIBSSH2_AGENT * agent) nogil except 1:
+    cdef c_ssh2.libssh2_agent_publickey *identity = NULL
+    cdef c_ssh2.libssh2_agent_publickey *prev = NULL
+    if c_ssh2.libssh2_agent_list_identities(agent) != 0:
+        clear_agent(agent)
+        with gil:
+            raise AgentListIdentitiesError(
+                "Failure requesting identities from agent")
+        return 1
+    while 1:
+        if auth_identity(_username, agent, &identity, prev) == 1:
+            with gil:
+                raise AgentAuthenticationError(
+                    "No identities match for user %s",
+                    _username)
+            return 1
+        if c_ssh2.libssh2_agent_userauth(
+                agent, _username, identity) == 0:
+            clear_agent(agent)
+            return 0
+        prev = identity
+    return 1
 
 
 cdef int auth_identity(const char *username,
@@ -30,10 +55,6 @@ cdef int auth_identity(const char *username,
         agent, identity, prev)
     if rc == 1:
         clear_agent(agent)
-        with gil:
-            raise AgentAuthenticationError(
-                "No identities match for user %s",
-                username)
     elif rc < 0:
         clear_agent(agent)
         with gil:
@@ -41,6 +62,27 @@ cdef int auth_identity(const char *username,
                 "Failure getting identity for user %s from agent",
                 username)
     return rc
+
+
+cdef c_ssh2.LIBSSH2_AGENT * agent_init(
+        c_ssh2.LIBSSH2_SESSION *_session) nogil except NULL:
+    cdef c_ssh2.LIBSSH2_AGENT *agent = c_ssh2.libssh2_agent_init(
+        _session)
+    if agent is NULL:
+        with gil:
+            raise AgentError("Error initialising agent")
+    return agent
+
+
+cdef c_ssh2.LIBSSH2_AGENT * init_connect_agent(
+        c_ssh2.LIBSSH2_SESSION *_session) nogil except NULL:
+    cdef c_ssh2.LIBSSH2_AGENT *agent
+    agent = c_ssh2.libssh2_agent_init(_session)
+    if c_ssh2.libssh2_agent_connect(agent) != 0:
+        c_ssh2.libssh2_agent_free(agent)
+        with gil:
+            raise AgentConnectionError("Unable to connect to agent")
+    return agent
 
 
 cdef void clear_agent(c_ssh2.LIBSSH2_AGENT *agent) nogil: