+++ tests/test_session.py
@@ -2,7 +2,9 @@ import os
 from unittest import skipUnless
 
 from .base_test import SSH2TestCase
-from ssh2.session import Session
+from ssh2.session import Session, LIBSSH2_HOSTKEY_HASH_MD5, \
+    LIBSSH2_HOSTKEY_HASH_SHA1
+from ssh2.exceptions import AuthenticationError, AgentAuthenticationError
 
 
 class SessionTestCase(SSH2TestCase):
@@ -30,6 +32,20 @@ class SessionTestCase(SSH2TestCase):
         self.assertEqual(self.session.get_blocking(), 0)
         self.session.set_blocking(1)
         self.assertEqual(self.session.get_blocking(), 1)
+
+    def test_failed_agent_auth(self):
+        self.assertRaises(AgentAuthenticationError,
+                          self.session.agent_auth, 'FAKE USER')
+
+    def test_failed_pkey_auth(self):
+        self.assertRaises(AuthenticationError,
+                          self.session.userauth_publickey_fromfile,
+                          'FAKE USER', self.user_pub_key, self.user_key,
+                          '')
+        self.assertRaises(AuthenticationError,
+                          self.session.userauth_publickey_fromfile,
+                          self.user, 'FAKE FILE', 'EVEN MORE FAKE FILE',
+                          '')
 
     @skipUnless(hasattr(Session, 'scp_recv64'),
                 "Function not supported by libssh2")
@@ -109,3 +125,9 @@ class SessionTestCase(SSH2TestCase):
         finally:
             os.unlink(remote_filename)
             os.unlink(to_copy)
+
+    def test_hostkey(self):
+        self.assertEqual(self._auth(), 0)
+        for _type in [LIBSSH2_HOSTKEY_HASH_MD5, LIBSSH2_HOSTKEY_HASH_SHA1]:
+            hostkey = self.session.hostkey_hash(_type)
+            self.assertTrue(len(hostkey) > 0)